* Added padding to crawl bitmap size. * Changed ImSoundId to signed type large enough to hold a pointer and hold the "midi flag" bit.   - This is needed because digital audio ids are really pointers. * Changed all sound ID references to ImSoundId. * Forced the iMuseSoundType enum to be unsigned. * Changed intptr_t to iptr. * Added iMuse digital sound files (still WIP).

+++ TheForceEngine/TFE_Jedi/IMuse/imConst.h
@@ -91,7 +91,7 @@ enum iMuseConst
 
 namespace TFE_Jedi
 {
-	typedef u32 ImSoundId;
+	typedef s64 ImSoundId;
 
 	extern const u8  c_midiMsgSize[];
 	extern const u32 c_channelMask[MIDI_CHANNEL_COUNT];
